+++ src/gt4py/next/program_processors/runners/dace_fieldview/utility.py
@@ -16,41 +16,13 @@ from typing import Any, Mapping
 
 import dace
 
-from gt4py.eve import codegen
-from gt4py.eve.codegen import FormatTemplate as as_fmt
 from gt4py.next.common import Connectivity, Dimension
 from gt4py.next.iterator import ir as itir
 from gt4py.next.iterator.ir_utils import common_pattern_matcher as cpm
-from gt4py.next.program_processors.runners.dace_fieldview.gtir_types import MATH_BUILTINS_MAPPING
+from gt4py.next.program_processors.runners.dace_fieldview.gtir_python_codegen import (
+    GTIRPythonCodegen,
+)
 from gt4py.next.type_system import type_specifications as ts
-
-
-class SymbolicTranslator(codegen.TemplatedGenerator):
-    SymRef = as_fmt("{id}")
-    Literal = as_fmt("{value}")
-
-    def _visit_deref(self, node: itir.FunCall) -> str:
-        assert len(node.args) == 1
-        if isinstance(node.args[0], itir.SymRef):
-            return self.visit(node.args[0])
-        raise NotImplementedError(f"Unexpected deref with arg type '{type(node.args[0])}'.")
-
-    def _visit_numeric_builtin(self, node: itir.FunCall) -> str:
-        assert isinstance(node.fun, itir.SymRef)
-        fmt = MATH_BUILTINS_MAPPING[str(node.fun.id)]
-        args = self.visit(node.args)
-        return fmt.format(*args)
-
-    def visit_FunCall(self, node: itir.FunCall) -> str:
-        if cpm.is_call_to(node, "deref"):
-            return self._visit_deref(node)
-        elif isinstance(node.fun, itir.SymRef):
-            builtin_name = str(node.fun.id)
-            if builtin_name in MATH_BUILTINS_MAPPING:
-                return self._visit_numeric_builtin(node)
-            else:
-                raise NotImplementedError(f"'{builtin_name}' not implemented.")
-        raise NotImplementedError(f"Unexpected 'FunCall' node ({node}).")
 
 
 def as_dace_type(type_: ts.ScalarType) -> dace.typeclass:
@@ -112,7 +84,7 @@ def get_domain(
 
 
 def get_symbolic_expr(node: itir.Expr) -> str:
-    return SymbolicTranslator().visit(node)
+    return GTIRPythonCodegen().visit(node)
 
 
 def unique_name(prefix: str) -> str:
